requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.0.1
email-validator==2.1.1
gunicorn==21.2.0
